Fix all Ruff code quality issues and restore strict checking rules
- Fix C408: Replace dict() calls with dict literals (4 files)
- Fix SIM105: Use contextlib.suppress instead of try-except-pass (2 files)
- Fix SIM102: Combine nested if statements (2 files)
- Fix SIM103: Return boolean expressions directly (2 files)
- Fix SIM110: Use any() instead of for loop (1 file)
- Fix RUF015: Use next(iter()) instead of list()[0] (1 file)
- Fix RUF012: Add ClassVar annotation for mutable class attributes (2 files)
- Fix RUF046: Remove unnecessary int() cast on round() result (2 files)
- Fix C416: Simplify dict/list comprehensions (2 files)
- Fix N806: Rename UPPER_CASE variables in functions to lowercase (5 files)
- Fix N818: Rename ExecutionFailure to ExecutionError (follows convention)
- Fix F841: Remove or mark unused variables with _ (4 files)
- Fix B007: Use _ for unused loop variables (1 file)

All 19 Ruff errors fixed. Restored strict checking rules in pyproject.toml.
Only keeping essential ignores: TC001-TC003 (FastAPI), RUF001-RUF003 (Greek letters).
All tests pass (942 passed, 1 skipped).

+++ app/database.py
@@ -323,7 +323,7 @@ class DuckDBStore:
                                 ):
                                     base_ctr += 0.008
                                 ctr = max(0.02, min(0.45, base_ctr + rng.uniform(-0.015, 0.015)))
-                                clicks = max(1, int(round(impressions * ctr)))
+                                clicks = max(1, round(impressions * ctr))
                                 recommendation_rows.append(
                                     (
                                         current_day,
